👌 IMPROVE: extract axios call
